refactor: Separate each constant into a file
