+++ .eslintrc
@@ -14,4 +14,3 @@
     "prettier/prettier": "warn"
   }
 }
-
