--- src/utils/Constants.js
@@ -1,19 +1,0 @@
-export const EventType = Object.freeze({
-  CLICK: 'click',
-  DOUBLE_CLICK: 'dblclick',
-  KEY_DOWN: 'keydown',
-});
-
-export const KeyType = Object.freeze({
-  ENTER: 'Enter',
-  ESC: 'Escape',
-});
-
-export const TextType = Object.freeze({
-  EMPTY: '',
-});
-
-export const TagName = Object.freeze({
-  A: 'A',
-  LIST: 'li',
-});
