+++ src/utils/EventType.js
@@ -1,0 +1,5 @@
+export const EventType = Object.freeze({
+  CLICK: 'click',
+  DOUBLE_CLICK: 'dblclick',
+  KEY_DOWN: 'keydown',
+});
