+++ src/utils/KeyType.js
@@ -1,0 +1,4 @@
+export const KeyType = Object.freeze({
+  ENTER: 'Enter',
+  ESC: 'Escape',
+});
