+++ src/utils/TagName.js
@@ -1,0 +1,4 @@
+export const TagName = Object.freeze({
+  A: 'A',
+  LIST: 'li',
+});
