+++ src/utils/TextType.js
@@ -1,0 +1,3 @@
+export const TextType = Object.freeze({
+  EMPTY: '',
+});
